Move miner to its own module
