+++ src/cryptonote_basic/CMakeLists.txt
@@ -33,7 +33,7 @@ add_library(cryptonote_basic
   cryptonote_format_utils.cpp
   difficulty.cpp
   hardfork.cpp
-  miner.cpp)
+        ../miner/miner.cpp)
 
 target_link_libraries(cryptonote_basic
   PRIVATE
