+++ src/cryptonote_core/cryptonote_core.h
@@ -47,7 +47,7 @@
 #include "service_node_voting.h"
 #include "service_node_list.h"
 #include "service_node_quorum_cop.h"
-#include "cryptonote_basic/miner.h"
+#include "miner/miner.h"
 #include "cryptonote_basic/connection_context.h"
 #include "warnings.h"
 #include "crypto/hash.h"
