+++ src/miner/miner.cpp
