+++ src/miner/miner.h
